fixed structures

+++ Models/PickListsLine.cs
@@ -5,6 +5,6 @@
         public int BaseObjectType { get; set; }
         public int OrderEntry {  get; set; }
         public int OrderRowID { get; set; }
-        public double ReleasedQuantity { get; set; }
+        public int ReleasedQuantity { get; set; }
     }
 }
